Bump wgpu from `0.13` to `0.14` (#320)
* feat(wgpu): upgrade wgpu from 13 to 14

* Address feedback

* Bump tao in `minimal-tao` example

* Bump fltk in `minimal-fltk` example

* Bump egui in `minimal-egui` example

* Bump MSRV to `1.65`

* Bump crates in `imgui-winit` example

* Delete `minimal-sdl2` example

* Fix clippy lints

* Update examples/minimal-egui/src/gui.rs

+++ examples/invaders/simple-invaders/src/loader.rs
@@ -67,7 +67,7 @@ fn load_pcx(pcx: &[u8]) -> CachedSprite {
         let mut buffer = Vec::new();
         buffer.resize_with(width * height, Default::default);
         for y in 0..height {
-            let a = y as usize * width;
+            let a = y * width;
             let b = a + width;
             reader.next_row_paletted(&mut buffer[a..b]).unwrap();
         }
